Sub Fields added
Added the ability for valueTypes to be json itself.  This makes it
wierd with Neo4J as then pretty much some entries have just a JSON
string as its value but I think this will play nice with DB's like
MongoDB and SQL and others, Neo4J is an outlier it database style so
I assumed it was going to be a little wierd.